test(lab/mega-menu): source e2e fixtures from shared cypress stories
